flaskapp/requirements.txt
flask
gunicorn
requests
flask-bootstrap
flask-wtf
pillow
tensorflow
keras